Add 'layers' to project config as list of strings
